Better handling of currency
current_from is always the point at which the record was updated
current_to is always the the current_from of the latest record

Essentially the current_to->current_from should flow.  Sure that the
messages arriving out of order will complicate this but it seems
sensible at the moment.