[fixes #28399055] Use triggers to maintain current tables
The various current views have been replaced by physical tables that are
maintained by triggers on the normal tables.  When a current row (one
with current_to NULL) or a deleted row (one with deleted_at not NULL) is
inserted into the normal tables the associated current table will be
updated.

Note that this now requires the column ordering to be the same between
these tables, i.e. a new column added to the normal table will require
the current table to be updated too.