Remove all unnecessary indexes & tighten column definitions
The current tables only require (at the moment) indexes on uuid (for the
trigger) and internal_id (for the hell of it).  Equally the normal
tables only need a unique compound index on uuid, current_from and
current_to.

Various common columns, like internal_id, can be constrained to not null
and uniqueness in some cases.