[#28956849] Implement deadlettering in AmqpConsumer
Any messages that fail to be processed, after being rejected once
already, will be sent to a dead letter exchange.  These failures may be
intermittent (due to DB deadlock issues, for example, when rebuild the
entire warehouse) or not, but putting them in another queue allows us to
retry them later.

Note that this can be done with RabbitMQ dead lettering support, but
that has been unreliable, killing the main queue (which is very bad), so
we're manually implementing this and it appears more stable.