[fixes #29989473] Maintain current_to properly
Because we've changed to a composite primary key of (uuid,current_to)
when updating the current_to we obviously change the primary key of the
record.  In theory this is actually OK, but the UPDATE statement uses
the new and, obviously, different current_to value; so no records
actually get updated (thanks ActiveRecord for not erroring!).  So we use
a specific SQL UPDATE statement to make the change.